produtoCpf/requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
bitrix24-rest==1.1.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
dialogflow==1.1.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.2
twilio==6.45.4
urllib3==1.25.10
validate-email==1.3
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
